requirements.txt
certifi==2018.11.29
chardet==3.0.4
convertdate==2.1.3
cycler==0.10.0
Cython==0.29.3
Django==2.1.5
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.1
ephem==3.7.6.0
fbprophet==0.4.post2
gevent==1.4.0
greenlet==0.4.15
holidays==0.9.9
idna==2.8
kiwisolver==1.0.1
lunardate==0.2.0
matplotlib==3.0.2
numpy==1.16.0
pandas==0.23.4
prometheus-client==0.5.0
PyMySQL==0.9.3
pyparsing==2.3.1
pystan==2.17.1.0
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
setuptools-git==1.2
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
vine==1.2.0
